Create first page comp w movie posters, added styling, started comp for movie details
