--- code/public/index.css
@@ -1,5 +1,0 @@
-body {
-    display: flex;
-    flex-direction: row;
-    flex-wrap: wrap;
-}
