+++ code/src/MovieDetails.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+export const Moviedetails = () => {
+    return (
+        <div>
+        halloj!
+    </div>
+
+    )
+    
+}
