--- code/src/components/moviePoster.css
@@ -1,5 +1,0 @@
-.movie {
-    display: flex;
-    flex-direction: column;
-    border:solid 2px green;
-}
